update after testing 3 type of running stored procedure with any return type but must determine it before create repos func.

+++ src/main/java/com/a08r/awcolabsapp/models/dto/LabsDailyTestDTO.java
@@ -11,7 +11,7 @@ public class LabsDailyTestDTO {
 
     private String testDate;
     private int labCode;
-    private double testValueAvg;
+    private float testValueAvg;
 }
 
 //public class LabsDailyTestDTO {
